🚚 [variables] Move module `variables` from `templates.domain`
