+++ tests/unit/conftest.py
@@ -4,7 +4,7 @@ import pytest
 from cutty.templates.domain.render import createrenderer
 from cutty.templates.domain.render import defaultrenderregistry
 from cutty.templates.domain.render import Renderer
-from cutty.templates.domain.variables import GenericVariable
+from cutty.variables.variables import GenericVariable
 
 
 @pytest.fixture
